better file structure and properly invalidating queries
