--- hooks/index.ts
@@ -1,4 +1,0 @@
-
-export { useHomepage } from './queries/use-homepage'
-
-export { queryKeys } from './queries/query-keys'
